requirements.txt
aiohttp
lxml_html_clean
Flask
gunicorn
werkzeug
aioshutil
aiofiles
anitopy
apscheduler
dnspython
motor
feedparser
psutil
pycountry
pymongo
pyrofork
python-dotenv
tgcrypto
html-telegraph-poster
uvloop
python-telegram-bot
ffmpeg-python
torrentp
